pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "FitsProcessor"
version = "0.1.0"
description = "A project for processing FITS files and generating catalogs."
authors = [
    { name = "Chaitanya Chawak", email = "chaitanya.chawak@cea.fr" }
]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "astropy>=5.0",
    "numpy>=1.21",
    "pyyaml>=6.0"
]

[project.urls]
repository = "https://github.com/ChaitanyaChawak/FitsProcessor"
